Move `forEachResponse` extension to `MultiStatusItemExtensions.kt`
